Update - The hosting slot/auction relationship is no longer mandatory. The slots can be created without references to any auction
git-svn-id: https://coder.topcoder.com/tcs/clients/cronos/components@72554 fb370eea-3af6-4597-97f7-f7400a59c12a

+++ ball/trunk/src/java/main/com/orpheus/game/server/handler/SponsorRegistrationHandler.java
@@ -202,7 +202,7 @@ public class SponsorRegistrationHandler implements Handler {
                         // This error is ignored as we have no other way to check if the file exists in FileUpload
                     }
                 }
-                domains.add(new DomainImpl(null, sponsorId.longValue(), request.getParameter("domain_" + i),
+                domains.add(new DomainImpl(null, sponsorId, request.getParameter("domain_" + i),
                                            null, (ImageInfo[]) images.toArray(new ImageInfoImpl[images.size()])));
             }
         }
